Refactor e-signature component: extract upload and choose modes into separate components
- Moved upload functionality to RdsESignatureUpload component, handling file selection and validation.
- Created RdsESignatureChoose component for selecting predefined signatures with validation for multiple selections.
- Updated main RdsCompESignature component to utilize the new mode components, simplifying the code structure.
- Removed unused state and handlers related to file upload and signature selection from the main component.
